+++ spec/slack/real_time/client_spec.rb
@@ -61,6 +61,10 @@ RSpec.describe Slack::RealTime::Client, vcr: { cassette_name: 'web/rtm_start' } 
           it 'sets team' do
             expect(client.team['domain']).to eq 'dblockdotorg'
           end
+          it 'sets teams' do
+            expect(client.teams.count).to eq 1
+            expect(client.teams.values.first).to eq client.team
+          end
           it 'sets self' do
             expect(client.self['id']).to eq 'U07518DTL'
           end